requirements.txt
torch>=1.9.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.62.0 